refactor: flatten backend — rename package ohsheet → backend
Removes the redundant backend/ohsheet/ nesting. The Python package is now
imported as `backend` directly, mirroring the polyglot repo's frontend/
naming.

Layout before:
  backend/
  ├── ohsheet/        ← package
  ├── tests/
  └── pyproject.toml

Layout after:
  pyproject.toml
  backend/            ← package
  tests/

Why pyproject.toml moves to the repo root: hatchling expects the package
directory to be a child of the project root, so once `backend/` itself is
the package, pyproject.toml has to live one level up.

Changes:
- Move backend/ohsheet/* up into backend/, move tests/ and pyproject.toml
  to the repo root.
- Rewrite all 19 Python files: `from ohsheet.X` → `from backend.X`.
- pyproject.toml: packages = ["backend"]; script entry = "backend.main:run"
  (CLI command name `ohsheet` and distribution name `ohsheet` are unchanged).
- Makefile: drop `cd $(BACKEND) &&` from backend / install / test targets;
  uvicorn now runs `backend.main:app` from the repo root.
- README: updated repo-layout and architecture trees.
- frontend/lib/api/models.dart: doc comment now references backend/.

Kept as `ohsheet`:
- pyproject.toml `name = "ohsheet"` — distribution / project name
- pyproject.toml `[project.scripts]` key — the CLI is still `ohsheet`
- backend/config.py `OHSHEET_` env-var prefix — user-facing config namespace

Verified:
- pytest 16/16 from the repo root
- flutter analyze clean
- live HTTP smoke test against `uvicorn backend.main:app` (health, upload,
  job submit, all 3 artifact downloads → 200)
