+++ backend/api/routes/stages.py
@@ -15,8 +15,8 @@ from typing import Annotated, Any, Awaitable, Callable
 from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel
 
-from ohsheet.api.deps import get_blob_store, get_runner
-from ohsheet.contracts import (
+from backend.api.deps import get_blob_store, get_runner
+from backend.contracts import (
     SCHEMA_VERSION,
     EngravedOutput,
     HumanizedPerformance,
@@ -26,8 +26,8 @@ from ohsheet.contracts import (
     TranscriptionResult,
     WorkerResponse,
 )
-from ohsheet.jobs.runner import PipelineRunner
-from ohsheet.storage.local import LocalBlobStore
+from backend.jobs.runner import PipelineRunner
+from backend.storage.local import LocalBlobStore
 
 router = APIRouter()
