+++ backend/jobs/__init__.py
@@ -1,0 +1,7 @@
+"""Async job execution + WebSocket pub/sub."""
+
+from backend.jobs.events import JobEvent, JobStatus
+from backend.jobs.manager import JobManager, JobRecord
+from backend.jobs.runner import PipelineRunner
+
+__all__ = ["JobEvent", "JobStatus", "JobManager", "JobRecord", "PipelineRunner"]
