+++ backend/jobs/manager.py
@@ -12,9 +12,9 @@ import uuid
 from dataclasses import dataclass, field
 from typing import Optional
 
-from ohsheet.contracts import EngravedOutput, InputBundle, PipelineConfig
-from ohsheet.jobs.events import JobEvent, JobStatus
-from ohsheet.jobs.runner import PipelineRunner
+from backend.contracts import EngravedOutput, InputBundle, PipelineConfig
+from backend.jobs.events import JobEvent, JobStatus
+from backend.jobs.runner import PipelineRunner
 
 log = logging.getLogger(__name__)
