--- backend/ohsheet/services/__init__.py
@@ -1,15 +1,0 @@
-"""Pipeline stage services. Stubs for now; wrappers around real workers later."""
-
-from ohsheet.services.arrange import ArrangeService
-from ohsheet.services.engrave import EngraveService
-from ohsheet.services.humanize import HumanizeService
-from ohsheet.services.ingest import IngestService
-from ohsheet.services.transcribe import TranscribeService
-
-__all__ = [
-    "IngestService",
-    "TranscribeService",
-    "ArrangeService",
-    "HumanizeService",
-    "EngraveService",
-]
