+++ backend/services/__init__.py
@@ -1,0 +1,15 @@
+"""Pipeline stage services. Stubs for now; wrappers around real workers later."""
+
+from backend.services.arrange import ArrangeService
+from backend.services.engrave import EngraveService
+from backend.services.humanize import HumanizeService
+from backend.services.ingest import IngestService
+from backend.services.transcribe import TranscribeService
+
+__all__ = [
+    "IngestService",
+    "TranscribeService",
+    "ArrangeService",
+    "HumanizeService",
+    "EngraveService",
+]
