+++ backend/services/humanize.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 import asyncio
 
-from ohsheet.contracts import (
+from backend.contracts import (
     SCHEMA_VERSION,
     ExpressionMap,
     ExpressiveNote,
