+++ backend/services/ingest.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 
 import asyncio
 
-from ohsheet.contracts import (
+from backend.contracts import (
     SCHEMA_VERSION,
     InputBundle,
     InputMetadata,
