+++ backend/services/transcribe.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 
 import asyncio
 
-from ohsheet.contracts import (
+from backend.contracts import (
     SCHEMA_VERSION,
     HarmonicAnalysis,
     InputBundle,
