+++ backend/storage/__init__.py
@@ -1,6 +1,6 @@
 """Blob storage abstraction (Claim-Check pattern)."""
 
-from ohsheet.storage.base import BlobStore
-from ohsheet.storage.local import LocalBlobStore
+from backend.storage.base import BlobStore
+from backend.storage.local import LocalBlobStore
 
 __all__ = ["BlobStore", "LocalBlobStore"]
